GCC 15 vs regex, qso browser   * Changed regex 'C' declarations to current standards.   * Commented out frequency2 list legacy code, causes incorrect assignment.   * Minor changes to mode parsing.   * Added qso_browser clear() function to prevent list doubling on     each frequency2 file reading.   * Modified code to prevent duplicate entries.

+++ src/include/rigsupport.h
@@ -42,7 +42,7 @@ extern Cserial rigio;
 extern void initOptionMenus();
 extern void clearList();
 extern void updateSelect();
-extern size_t addtoList(long val);
+extern long addtoList(long long val);
 extern void build_frequencies2_list();
 extern void qso_movFreq(Fl_Widget* w, void*);
 extern int	cb_qso_opMODE();
